+++ ndsi-engine/src/main/java/br/com/abril/nds/integracao/ems0107/inbound/EMS0107Input.java
@@ -2,6 +2,7 @@ package br.com.abril.nds.integracao.ems0107.inbound;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -41,7 +42,7 @@ public class EMS0107Input implements Serializable {
 	private Long codigoCota;
 
 	@Column
-	private BigDecimal quantidadeReparte;
+	private BigInteger quantidadeReparte;
 
 	@Column
 	private String repartePDV;
@@ -80,11 +81,11 @@ public class EMS0107Input implements Serializable {
 
 	@Field(offset = 17, length = 8)
 	@FixedFormatDecimal(decimals = 2, useDecimalDelimiter = false)
-	public BigDecimal getQuantidadeReparte() {
+	public BigInteger getQuantidadeReparte() {
 		return quantidadeReparte;
 	}
 
-	public void setQuantidadeReparte(BigDecimal quantidadeReparte) {
+	public void setQuantidadeReparte(BigInteger quantidadeReparte) {
 		this.quantidadeReparte = quantidadeReparte;
 	}
